# docs/swagger.yaml
basePath: /
definitions:
  error.Error:
    properties:
      error_code:
        type: string
      message:
        type: string
    type: object
...
  model.ApplicationView:
    properties:
      applications:
        items:
          $ref: '#/definitions/model.ViewApplicationInfo'
        type: array
      msg:
        type: string
      num:
        type: integer
    type: object
...
  model.RemindBox:
    properties:
      content:
        items:
          $ref: '#/definitions/model.ReminderInfo'
        type: array
      msg:
        type: string
      num:
        type: integer
    type: object
  model.RemindEx:
    properties:
      existence:
        type: boolean
      msg:
        type: string
    type: object
  model.ReminderInfo:
    properties:
      college:
        type: string
      confirm_time:
        type: string
      contact_way:
        type: string
      contact_way_type:
        type: string
      gender:
        type: string
      grade:
        type: string
      nick_name:
        type: string
      red_point:
        type: boolean
      requirement_id:
        type: integer
      status:
        type: integer
      title:
        type: string
    type: object
...
  model.Res:
    properties:
      msg:
        type: string
    type: object
...
  model.ViewApplicationInfo:
    properties:
      application_id:
        type: integer
      college:
        type: string
      contact_way:
        type: string
      contact_way_type:
        type: string
      gender:
        type: string
      grade:
        type: string
      red_point:
        type: boolean
      requirements_id:
        type: integer
      send_time:
        type: string
      sender_nickname:
        type: string
      title:
        type: string
    type: object
...
host: 47.97.74.180:9090
info:
  contact: {}
  description: 孤独星球
  license: {}
  title: lonely planet
  version: "1.0"
paths:
  /application/solve/{application_id}/:
    put:
      consumes:
      - application/json
      description: 根据请求Id来处理请求，id别的api会给出，通过状态status来处理请求，2为接受，3为拒绝
      parameters:
      - description: 状态
        in: query
        name: status
        required: true
        type: string
      - description: 请求id
        in: path
        name: application_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"success"}/{"msg":"需求已经被删除了!"}/{"msg":"已经处理过了!"}'
          schema:
            $ref: '#/definitions/model.Res'
        "400":
          description: '{"error_code":"00001", "message":"Fail."} or {"error_code":"00002",
            "message":"Lack Param Or Param Not Satisfiable."}'
          schema:
            $ref: '#/definitions/error.Error'
        "401":
          description: '{"error_code":"10001", "message":"Token Invalid."} 身份认证失败
            重新登录'
          schema:
            $ref: '#/definitions/error.Error'
        "500":
          description: '{"error_code":"30001", "message":"Fail."} 失败'
          schema:
            $ref: '#/definitions/error.Error'
      summary: 处理请求
      tags:
      - application
  /application/unsolve/:
    get:
      consumes:
      - application/json
      description: 点击申请提醒　查看所有待确认和接受的申请
      parameters:
      - description: 每页数量
        in: query
        name: limit
        required: true
        type: string
      - description: 当前请求页数，从0开始
        in: query
        name: page
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"success", "num":数量, "applications":数组，其中包含每个的id}，其中red_point字段是用来表示是否未读，即单条信息是否显示小红点'
          schema:
            $ref: '#/definitions/model.ApplicationView'
        "400":
          description: '{"error_code":"00001", "message":"Fail."} or {"error_code":"00002",
            "message":"Lack Param Or Param Not Satisfiable."}'
          schema:
            $ref: '#/definitions/error.Error'
        "401":
          description: '{"error_code":"10001", "message":"Token Invalid."} 身份认证失败
            重新登录'
          schema:
            $ref: '#/definitions/error.Error'
        "500":
          description: '{"error_code":"30001", "message":"Fail."} 失败'
          schema:
            $ref: '#/definitions/error.Error'
      summary: 申请提醒
      tags:
      - application
  /application/view_info/{application_id}/:
    post:
      consumes:
      - application/json
      description: 更新申收件人阅读状态, 需要在用户在申请提醒中点击需求或者点击小眼睛或者直接处理请求的同时, 通过请求这条来更新状态, 如果可以希望可以根据是否已读来判断是否进行此次请求以减少请求次数
      parameters:
      - description: 用户已查看的申请id, 在别的api中给出
        in: path
        name: application_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"success"} 成功'
          schema:
            $ref: '#/definitions/model.Res'
        "400":
          description: '{"error_code":"00001", "message":"Fail."} or {"error_code":"00002",
            "message":"Lack Param Or Param Not Satisfiable."}'
          schema:
            $ref: '#/definitions/error.Error'
        "401":
          description: '{"error_code":"10001", "message":"Token Invalid."} 身份认证失败
            重新登录'
          schema:
            $ref: '#/definitions/error.Error'
        "500":
          description: '{"error_code":"30001", "message":"Fail."} 失败'
          schema:
            $ref: '#/definitions/error.Error'
      summary: 更新收件人阅读状态
      tags:
      - remind
...
  /remind/day/remindbox/:
    get:
      consumes:
      - application/json
      description: 查看回复提醒
      parameters:
      - description: 每页数量
        in: query
        name: limit
        required: true
        type: string
      - description: 当前请求页数，从0开始
        in: query
        name: page
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"success", "num":数量, "content":数组，其中包含每个的id， 其中confirm是用来判断显示的内容是否带有小眼睛图标，
            2为接受，3为拒绝，其中red_point字段是用来表示是否未读，即单条信息是否显示小红点}'
          schema:
            $ref: '#/definitions/model.RemindBox'
        "400":
          description: '{"error_code":"00001", "message":"Fail."} or {"error_code":"00002",
            "message":"Lack Param Or Param Not Satisfiable."}'
          schema:
            $ref: '#/definitions/error.Error'
        "401":
          description: '{"error_code":"10001", "message":"Token Invalid."} 身份认证失败
            重新登录'
          schema:
            $ref: '#/definitions/error.Error'
        "500":
          description: '{"error_code":"30001", "message":"Fail."} 失败'
          schema:
            $ref: '#/definitions/error.Error'
      summary: 回复提醒
      tags:
      - remind
  /remind/day/remindbox/info/{application_id}/:
    post:
      consumes:
      - application/json
      description: 更新申请人阅读状态, 需要在用户在回复提醒中点击需求或者点击小眼睛的同时, 通过请求这条来更新状态, 如果可以希望可以根据是否已读来判断是否进行此次请求以减少请求次数
      parameters:
      - description: 用户已查看的申请id, 在别的api中给出
        in: path
        name: application_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"success"} 成功'
          schema:
            $ref: '#/definitions/model.Res'
        "400":
          description: '{"error_code":"00001", "message":"Fail."} or {"error_code":"00002",
            "message":"Lack Param Or Param Not Satisfiable."}'
          schema:
            $ref: '#/definitions/error.Error'
        "401":
          description: '{"error_code":"10001", "message":"Token Invalid."} 身份认证失败
            重新登录'
          schema:
            $ref: '#/definitions/error.Error'
        "500":
          description: '{"error_code":"30001", "message":"Fail."} 失败'
          schema:
            $ref: '#/definitions/error.Error'
      summary: 更新申请人阅读状态
      tags:
      - remind
  /remind/day/remindbox/status/:
    get:
      consumes:
      - application/json
      description: 根据返回值来判断是否显示提醒的小红点
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"success", "existence":"true/false"}'
          schema:
            $ref: '#/definitions/model.RemindEx'
        "400":
          description: '{"error_code":"00001", "message":"Fail."} or {"error_code":"00002",
            "message":"Lack Param Or Param Not Satisfiable."}'
          schema:
            $ref: '#/definitions/error.Error'
        "401":
          description: '{"error_code":"10001", "message":"Token Invalid."} 身份认证失败
            重新登录'
          schema:
            $ref: '#/definitions/error.Error'
        "500":
          description: '{"error_code":"30001", "message":"Fail."} 失败'
          schema:
            $ref: '#/definitions/error.Error'
      summary: 查询是否显示小红点
      tags:
      - remind
...
swagger: "2.0"